requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
